Update loglik.py to handle padding, see extended description for more details
Not sure if this is the most efficient way to do this, but I find the last non-zero span, take that index, and just implemented the exact logic as the original loglik function. I also have an extra check in case someone throws a padding [1.0, 0.0] in between non-zero span entries but maybe this isn't necessary if we assume the users would just have valid consecutive nonzero spans.